pyproject.toml
[tool.poetry]
name = "gps_service_api"
version = "0.1.0"
description = ""
authors = ["VitalyDubovets <dubovetsvitaliy@gmail.com>"]

[tool.poetry.dependencies]
python = "3.6.9"
django = "3.0.4"
djangorestframework = "3.11.0"
django-filter = "2.2.0"
django-redis = "4.11.0"
psycopg2-binary = "2.8.4"
python-dotenv = "0.12.0"
geopy = "^1.21.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
